Fixed KVMetadata.Version when over max_versions
TIL Vault will start purging versions of a secret over a set maximum.
But not just the secret - its record of existing as well (unlike when a
secret is destroyed). This threw off the naive indexing code that I had
there previously. Now it calculates the index the secret is at by
offsetting from the first returned version.