Bump tests to 1.0.0. Fix gen root.
Generate root changed how it reacts to you trying to give it an OTP of
the wrong length (16 bytes), and also, the token returned is no longer
on a 4-byte offset because of the added two-byte prefix, and therefore
the base64 encoding needs padding... which is trimmed by the server. So
we needed to add our own padding before trying to decode it.