Add TokenIsValid() and handle 403 in isKVv2Mount
Older versions of Vault will 403 if you hit the sys/internal/ui/mounts
endpoint but your token does not have read rights on sys/*. Previously,
this only handled the case of a 404 (which occurs when you have rights
to the sys mount). Now this can properly be detected as the mount being
v1. The TokenIsValid() function was made to help with this, as we still
want to error if somebody just gave us a bad token (because the mount
could actually be v2 in that case!). This still fails if the version of
Vault is earlier than 0.9, the token has no read rights on sys, and the
policy admin has revoked the _default_ right to inspect one's own token.
If all of these apply to you, I'm sorry.