Special error message for getting HTML response
The old default message when the JSON stream failed to parse of
something like "expected '<' in byte stream" confused a bunch of people.
Now, if the stream could not be parsed as JSON, it will give a more
descriptive error in the case that the Content-Type header returned did
not claim to be sending back JSON anyway. This usually happens if you're
targeting a non-Vault, or something along the way is stopping you before
you get to Vault, like nginx.