check for seal status when generating root token
Previously, we just relied on the Vault server telling us if the server
is sealed or uninitialized, which worked until Vault 1.4 and 1.5 had a
bug where the server would crash if you tried to do this. Now we short
circuit the logic if we determine beforehand that the vault is sealed or
uninitialized.